added state dicts to states.py and moved states.py to universals package instead of top level directory
